[Tests] Simple project that consumes OpenAssetIO
Add a new CMake project that consumes OpenAssetIO, either as subproject
or a package. CTest tests then check each component of OpenAssetIO can
be imported/included/linked.

Note that Python tests in subproject mode are currently disabled, due
to the complexities of marshalling Python source and binary modules.
This was a pragmatic choice, since the main focus of development is
toward a redistributable package, rather than use as a subproject, but
will hopefully be supported at a later date.

In subproject mode `add_subdirectory` is used to bring in OpenAssetIO
via a git submodule, which must be initialised externally.

In package mode `find_package` is used to bring in OpenAssetIO from an
installation, which must be made available to CMake's search paths
externally.

Tests are extremely minimal and do not exercise OpenAssetIO itself. The
intention is just to test basic integration in a consuming project for
issues with linkage etc.