[Build] Support linking Python as a static library
If `libpython` is linked in to the `test.python.bridge` executable as a
static library, then by default none of `libpython`'s symbols will be
visible to Python extension modules, causing "undefined reference"
errors when importing. This is simply the default behaviour for
executable targets.

So instruct CMake to make the executable's symbols visible.

Signed-off-by: David Feltell <david.feltell@foundry.com>