[Build] Set a default build type of Release
The default is toolchain-specific (i.e. CMake generator-specific) and
this can cause unexpected errors.

E.g. if we've installed Python as a Release library, but this project is
built as a Debug library, then on Windows we'll try to link to
`python39_d`, i.e. the debug lib, which may not exist.

We could enforce that `CMAKE_BUILD_TYPE` is always set externally,
however if we forget then builds might just work, but then fail
unexpectedly if we swap the toolchain.

So better to just be consistent and default to Release in all cases,
unless explicitly overridden.

Signed-off-by: David Feltell <david.feltell@foundry.com>