[CI] Pin actions/checkout in dependabot
The latest version of actions/checkout bumps its nodejs dependency to
v20, which requires a newer glibc version than is available in the
ASWF CY22 Docker image.

Ultimately we aim to use GitHub release artifacts, negating the need for
Docker (i.e. making a throwaway build of dependencies). So in the
meantime, just ignore major version updates of actions/checkout.

Signed-off-by: David Feltell <david.feltell@foundry.com>